requirements.txt
pandas==1.0.3
xlrd==1.2.0
attrs==19.3.0
path==13.1.0
openpyxl==3.0.3
pdfminer
PyYAML==5.3.1
xlsxwriter